Исправил некоторые варнинги после добавления флагов

+++ print.h
@@ -5,7 +5,7 @@
 #include "Stack.h"
 
 #define print_tabs(num_tab)                                     \
-    for(int i = 1; i <= num_tab; i++) fprintf(tree_file, "  ");
+    for (size_t i = 1; i <= num_tab; i++) fprintf(tree_file, "  ");
 
 void print_node_to_file(const node* now_node, size_t num_layer, FILE* tree_file);
 void print_definition  (elem_t name,    stack* name_stk,    stack* answer_stk);
